+++ src/net/xfr/mod.rs
@@ -4,5 +4,5 @@
 )]
 // #![warn(missing_docs)]
 // #![warn(clippy::missing_docs_in_private_items)]
-//! Zone transfer related functionality.
-pub mod processing;
+//! XFR protocol related functionality.
+pub mod protocol;
